requirements.txt
pip
wheel
setuptools 
virtualenv
kivy
docutils 
pygments
Cython==0.29.10
pillow
wifi
RPi.GPIO
asyncio
azure.iot.device
azure.identity
azure.keyvault.certificates
azure.keyvault.secrets
azure.keyvault.keys
bluepy